finished with layout for now - will need to update modals and get API requests set up
