+++ src/Components/ChapterPage/ChapterPage.js
@@ -1,13 +1,13 @@
 import React, { useState } from 'react'
 import Connections from './Connections'
+import ChapterMembers from "./ChapterMembers"
 import './ChapterPage.scss'
-import { Avatar } from '@material-ui/core' 
-import { AvatarGroup } from '@material-ui/lab';
+
 import kev from '../../Assets/Imgs/kev.jpg'
 import NavBar from '../MapComponents/Navbar'
 import { ToastProvider } from "react-toast-notifications";
 
-
+import headerImg from '../../Assets/Imgs/chapter.jpg'
 import pictureOne from '../../Assets/Imgs/Arash1.jpeg';
 import pictureTwo from '../../Assets/Imgs/Shawn.jpeg';
 import pictureThree from '../../Assets/Imgs/david.jpeg';
@@ -15,7 +15,7 @@ import pictureFour from '../../Assets/Imgs/Daniel.jpeg';
 import pictureFive from '../../Assets/Imgs/Ramonta.jpeg';
 import pictureSix from '../../Assets/Imgs/Will.jpeg';
 import { makeStyles } from '@material-ui/core/styles';
-import { withStyles } from '@material-ui/styles';
+
 
 const members = [
   {
@@ -53,25 +53,6 @@ const members = [
   }
 ]
 
-const LeaderAvatar = withStyles({
-  root: {
-    width: "25%",
-    height: "auto",
-    margin: "0 auto"
-  },
-  img: {
-    boxShadow: "0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)"
-  }
-})(Avatar)
-
-const VolunteerAvatar = withStyles({
-  root: {
-    width: "15%",
-    height: "auto",
-    margin: "0 auto"
-  },
-})(Avatar)
-
 
 const ChapterPage = () => {
   const [ chapterInfo, setChapterInfo ] = useState({
@@ -88,9 +69,6 @@ const ChapterPage = () => {
   
   return (
     <div className="chapter-page-container">
-    {/* <ToastProvider>
-      <NavBar/>
-    </ToastProvider> */}
       <div className="header-img"></div>
       <div className="inner-container">
         <h1 className="chapter-name">Seattle Chapter</h1>
@@ -113,20 +91,8 @@ const ChapterPage = () => {
           Join Chapter
         </button>
         <Connections />
-        <div className="chapter-leader">
-          <h2>Chapter Leader</h2>
-          <LeaderAvatar src={chapterInfo.leaderImg} />
-          <h2>{chapterInfo.chapterLeader}</h2>
-          <p>We are on the frontlines of COVID-19, reconnecting our homeless neighbors with their loved ones… and us.</p>
-        </div>
-        <div className="chapter-volunteers">
-          <h2>Volunteers</h2>
-          <AvatarGroup spacing="large" max={6}>
-            {chapterInfo.chapterMembers.map(el => {
-              return <VolunteerAvatar alt={el.name} src={el.photo} />
-            })}
-          </AvatarGroup>
-        </div>
+        <ChapterMembers chapterInfo={chapterInfo} />
+        {/* Should separate out chapter leader section as sep component */}
       </div>
     </div>
   )
