+++ src/Components/ChapterPage/ChapterPage.scss
@@ -4,7 +4,7 @@
   background-size: cover;
   background-position: right 50% bottom 80%;
   width: 100%;
-  height: 40vh;
+  height: 35vh;
 }
 
 p, h1, h2, h3, h4, button {
@@ -21,7 +21,7 @@ p, h1, h2, h3, h4, button {
     .chapter-name {
       font-size: 2.5rem;
       text-align: center;
-      padding: 3% 0;
+      padding: 2% 0;
       border-bottom: 5px solid #C0C0C0;
       font-weight: 900;
     }
@@ -35,7 +35,7 @@ p, h1, h2, h3, h4, button {
       
       p {
         text-align: center;
-        font-size: 2.5rem;
+        font-size: 2rem;
         font-weight: 800;    
       }
     }
@@ -43,37 +43,54 @@ p, h1, h2, h3, h4, button {
     .hero-p {
       padding: 2%;
       width: 100%;
-      font-size: 2rem;
+      font-size: 1.75rem;
     }
+
+    .members-header {
+        font-size: 2.5rem;
+        font-weight: 800;    
+        padding: 2% 0;
+        margin-bottom: 2%;
+      }
+
 
     .chapter-leader {
       display: flex;
       flex-direction: column;
-      justify-content: center;
+      width: 70%;
 
       h2 {
         text-align: center;
-        font-size: 2.5rem;
+        font-size: 2rem;
         font-weight: 800;    
         padding: 2% 0;
       }
 
-      p {
-        padding: 2%;
+      .bio {
         width: 100%;
-        font-size: 2rem;
+
+        h3 {
+          font-size: 2rem;
+        }
+
+        p {
+          padding: 2%;
+          width: 100%;
+          font-size: 1.75rem;
+          text-align: center;
+        }
       }
     }
 
     .chapter-volunteers {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
+      width: 30%;
 
       h2 {
         padding: 2%;
-        font-size: 2.5rem;
+        font-size: 2rem;
         font-weight: 800;
+        text-align: center;
+        margin-bottom: 2%;
       }
     }
   } 
@@ -81,11 +98,11 @@ p, h1, h2, h3, h4, button {
 
 .join-button {
   display: block;
-  margin: 2% auto;
+  margin: 1% auto;
   font-weight: 600;
-  width: 33.3%;
+  width: 30%;
   border-radius: 7px;
-  padding: 1.5% 0%;
+  padding: 1.0% 0%;
   color: #696969;
   font-size: 2rem;
   border: 2px solid #696969;
